requirements.txt
transformers
google-generativeai
python-dotenv
spotipy